+++ services/audio_service/client/src/audio_service_client.cpp
@@ -2407,7 +2407,7 @@ int32_t AudioServiceClient::GetCurrentPosition(uint64_t &framePosition, uint64_t
 
     timespec tm {};
     clock_gettime(CLOCK_MONOTONIC, &tm);
-    timestamp = tm.tv_sec * AUDIO_S_TO_NS + tm.tv_nsec;
+    timestamp = static_cast<uint64_t>(tm.tv_sec) * AUDIO_S_TO_NS + static_cast<uint64_t>(tm.tv_nsec);
 
     return SUCCESS;
 }
@@ -3512,7 +3512,7 @@ void AudioServiceClient::ProcessEvent(const AppExecFwk::InnerEvent::Pointer &eve
 
         // RenderMarkReach
         case RENDERER_MARK_REACHED_REQUEST:
-            mFrameMarkPosition = event->GetParam();
+            mFrameMarkPosition = static_cast<uint64_t>(event->GetParam());
             HandleRenderMarkReachedEvent(mFrameMarkPosition);
             break;
         case SET_RENDERER_MARK_REACHED_REQUEST:
@@ -3525,7 +3525,7 @@ void AudioServiceClient::ProcessEvent(const AppExecFwk::InnerEvent::Pointer &eve
 
         // RenderPeriodReach
         case RENDERER_PERIOD_REACHED_REQUEST:
-            mFramePeriodNumber = event->GetParam();
+            mFramePeriodNumber = static_cast<uint64_t>(event->GetParam());
             HandleRenderPeriodReachedEvent(mFramePeriodNumber);
             break;
         case SET_RENDERER_PERIOD_REACHED_REQUEST:
@@ -3538,7 +3538,7 @@ void AudioServiceClient::ProcessEvent(const AppExecFwk::InnerEvent::Pointer &eve
 
         // CapturerMarkReach
         case CAPTURER_MARK_REACHED_REQUEST:
-            mFrameMarkPosition = event->GetParam();
+            mFrameMarkPosition = static_cast<uint64_t>(event->GetParam());
             HandleCapturerMarkReachedEvent(mFrameMarkPosition);
             break;
         case SET_CAPTURER_MARK_REACHED_REQUEST:
@@ -3551,7 +3551,7 @@ void AudioServiceClient::ProcessEvent(const AppExecFwk::InnerEvent::Pointer &eve
 
         // CapturerPeriodReach
         case CAPTURER_PERIOD_REACHED_REQUEST:
-            mFramePeriodNumber = event->GetParam();
+            mFramePeriodNumber = static_cast<uint64_t>(event->GetParam());
             HandleCapturerPeriodReachedEvent(mFramePeriodNumber);
             break;
         case SET_CAPTURER_PERIOD_REACHED_REQUEST:
